LayoutTests/imported/w3c/web-platform-tests/encoding/legacy-mb-schinese/gb18030/gb18030-encoder.html

<!doctype html>
<meta charset=gb18030>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=resources/ranges.js></script>
<script>
 const encode = (input, output, desc) => {
   test(function() {
     const a = document.createElement("a"); // <a> uses document encoding for URL's query
     a.href = "https://example.com/?" + input;
     assert_equals(a.search.substr(1), output); // remove leading "?"
   }, "gb18030 encoder: " + desc);
 }

 encode("s", "s", "very basic");
 encode("\u20AC", "%A2%E3", "Euro");
 encode("\u4E02", "%81@", "character");
 encode("\uE4C6", "%A1@", "PUA");
 encode("\uE4C5", "%FE%FE", "PUA #2");
 encode("\uE5E5", "%26%2358853%3B", "PUA #3");
 encode("\ud83d\udca9", "%949%DA3", "poo");
 encode("\uE7C7", "%815%F47", "Ranges pointer special case");
 encode("\uE7C8", "%836%C80", "legacy ICU special case 1");
 encode("\u2026", "%A1%AD", "legacy ICU special case 2");
 encode("\uFF5E", "%A1%AB", "legacy ICU special case 3");

 encode("\uFE10", "%A6%D9", "GB18030-2022 1");
 encode("\uFE12", "%A6%DA", "GB18030-2022 2");
 encode("\uFE11", "%A6%DB", "GB18030-2022 3");
 encode("\uFE13", "%A6%DC", "GB18030-2022 4");
 encode("\uFE14", "%A6%DD", "GB18030-2022 5");
 encode("\uFE15", "%A6%DE", "GB18030-2022 6");
 encode("\uFE16", "%A6%DF", "GB18030-2022 7");
 encode("\uFE17", "%A6%EC", "GB18030-2022 8");
 encode("\uFE18", "%A6%ED", "GB18030-2022 9");
 encode("\uFE19", "%A6%F3", "GB18030-2022 10");
 encode("\u9FB4", "%FEY", "GB18030-2022 11");
 encode("\u9FB5", "%FEa", "GB18030-2022 12");
 encode("\u9FB6", "%FEf", "GB18030-2022 13");
 encode("\u9FB7", "%FEg", "GB18030-2022 14");
 encode("\u9FB8", "%FEm", "GB18030-2022 15");
 encode("\u9FB9", "%FE~", "GB18030-2022 16");
 encode("\u9FBA", "%FE%90", "GB18030-2022 17");
 encode("\u9FBB", "%FE%A0", "GB18030-2022 18");

 encode("\uE78D", "%836%CB2", "GB18030-2005 1");
 encode("\uE78E", "%836%CB3", "GB18030-2005 2");
 encode("\uE78F", "%836%CB4", "GB18030-2005 3");
 encode("\uE790", "%836%CB5", "GB18030-2005 4");
 encode("\uE791", "%836%CB6", "GB18030-2005 5");
 encode("\uE792", "%836%CB7", "GB18030-2005 6");
 encode("\uE793", "%836%CB8", "GB18030-2005 7");
 encode("\uE794", "%836%CB9", "GB18030-2005 8");
 encode("\uE795", "%836%CC0", "GB18030-2005 9");
 encode("\uE796", "%836%CC1", "GB18030-2005 10");
 encode("\uE81E", "%836%CA0", "GB18030-2005 11");
 encode("\uE826", "%836%CA7", "GB18030-2005 12");
 encode("\uE82B", "%836%CB1", "GB18030-2005 13");
 encode("\uE82C", "%836%CB2", "GB18030-2005 14");
 encode("\uE832", "%836%CB6", "GB18030-2005 15");
 encode("\uE843", "%836%CD0", "GB18030-2005 16");
 encode("\uE854", "%836%CE6", "GB18030-2005 17");
 encode("\uE864", "%836%D00", "GB18030-2005 18");

 const upperCaseNibble = x => {
   return Math.floor(x).toString(16).toUpperCase();
 }

 const encodePointer = pointer => {
   const firstByte = Math.floor(pointer / 12600) + 0x81;
   const thirdByte = Math.floor((pointer % 1260) / 10) + 0x81;
   return "%"
     + upperCaseNibble(firstByte / 16)
     + upperCaseNibble(firstByte % 16)
     + String.fromCharCode(Math.floor((pointer % 12600) / 1260) + 0x30)
     + "%"
     + upperCaseNibble(thirdByte / 16)
     + upperCaseNibble(thirdByte % 16)
     + String.fromCharCode(pointer % 10 + 0x30);
 }

 let i = 0;
 for (const range of ranges) {
   encode(range[1], encodePointer(range[0]), "range " + i++);
 }
</script>
